pyproject.toml
[project]
name = "omicidx-gh-etl"
version = "0.1.0"
description = ""
authors = [
    {name = "Sean Davis", email = "seandavi@gmail.com"},
]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "universal-pathlib >=0.2.2,<0.3.0",
    "rich >=13.7.1,<14.0.0",
    "s3fs >=2023.0.0,<2024.0.0",
    "orjson >=3.10.3,<4.0.0",
    "pubmed-parser @ git+https://github.com/seandavi/pubmed_parser",
    "gcsfs >=2023.0.0,<2024.0.0",
    "pydantic-settings >=2.3.3,<3.0.0",
    "clickhouse-connect >=0.7.12,<0.8.0",
    "google-cloud-bigquery >=3.24.0,<4.0.0",
    "pandas >=2.2.2,<3.0.0",
    "prefect>=3.4.2",
    "httpx>=0.28.1",
    "trio>=0.30.0",
    "python-dotenv[cli]>=1.1.0",
    "omicidx>=1.15.0",
    "duckdb>=1.3.0",
    "ibis-framework[bigquery,duckdb]>=10.5.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel >=6.29.4,<7.0.0",
]

[build-system]
requires = ["poetry-core>=1.5.0"] # Ensures poetry-core reads from [project]
build-backend = "poetry.core.masonry.api"

# This minimal [tool.poetry] section tells poetry-core where to find the local package(s).
# All other metadata and dependencies are now in the [project] section.
[tool.poetry]
packages = [{include = "omicidx_etl"}]
